Fix rounding bug for ex-time

+++ libs/sum_ms.php
@@ -1,6 +1,6 @@
 <?php
 
-$ms = intval($argv[1]??0);
-$total = intval($argv[2]??0);
+$ms = floatval($argv[1]??0);
+$total = floatval($argv[2]??0);
 
 echo $ms + $total;
